requirements/prod.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
python-decouple==3.8